add TaskController

+++ app/Http/Controllers/TasksController.php
@@ -98,9 +98,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.edit', [
+        if(\Auth::id() === $task->user_id){
+            return view('tasks.edit', [
             'task' => $task,
             ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -134,7 +138,12 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        
+        if(\Auth::id() === $task->user_id){
+            $task->delete();
+        } else { 
+            return redirect('/');
+        }
         
         return redirect('/');
     }
